Split server list logic into its own LogicObject.
